requirements.txt
anyio==3.6.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
et-xmlfile==1.1.0
Flask==2.1.2
gevent==21.12.0
gitdb==4.0.9
GitPython==3.1.27
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.21.6
openpyxl==3.0.9
pandas==1.3.5
pycparser==2.21
python-dateutil==2.8.2
python-gitlab==3.4.0
pytz==2022.1
requests==2.27.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
xmindparser==1.0.9
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
